Lubuntu LXQt wiki more canonical than daily PPA (#22)
* Lubuntu LXQt wiki more canonical than daily PPA

The Lubuntu daily PPA is not the recommended way of installing LXQt in Ubuntu. At this point, the metapackage doesn't function, among other problems. As of Xenial, we have had lxqt packages in the repos, so there's no reason not to recommend that instead. Additionally, more information about the future of LXQt in Lubuntu is available here. 

Oh, and don't worry. I keep saying Lubuntu because we're the part of the Ubuntu family that cares most about the future of LXDE/LXQt. You can just as easily install in Ubuntu. ☺

* changed title

per @agaida's suggestion, changed the title of the Ubuntu image along with the change to the link.